% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_RStudio.R
\name{config_workspace}
\alias{config_workspace}
\title{Configure an RStudio Workspace in the Research Cloud}
\usage{
config_workspace()
}
\description{
R script for configuring GitHub and iRODS connections from within an RStudio session.
Set a GitHub PAT
Install icomcands, create an irods_environment
Check connection
}
