% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_ws_project}
\alias{create_ws_project}
\title{Create Project in R Workspace}
\usage{
create_ws_project(project_name = NULL, git = TRUE, github = TRUE)
}
\arguments{
\item{project_name}{Name (string) of the project}

\item{git}{Do we want version control (default TRUE)}

\item{github}{Do we want an upstream/origin (default TRUE)}
}
\description{
Create Project in R Workspace
}
