% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_iRODS.R
\name{get_filenames_irsync}
\alias{get_filenames_irsync}
\title{Get Filenames from irsync list}
\usage{
get_filenames_irsync(irsync_list)
}
\arguments{
\item{irsync_list}{Output text (vector of strings) of \emph{irsync -l}}
}
\description{
Get Filenames from irsync list
}
\details{
Each line (except first) has three parts seperated by 3 spaces: filename, byte count, and ?
}
