% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_iRODS.R
\name{install_icommands}
\alias{install_icommands}
\title{Install irods icommands}
\usage{
install_icommands(out_file = "ii.out", err_file = "ii.err")
}
\arguments{
\item{out_file}{Basename of file in which to write output of installation script}

\item{err_file}{Idem for errors.}
}
\description{
Install irods icommands
}
\details{
The output and error files are stored in the working directory
}
\note{
Only works for Ubuntu x.x
}
