% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_iRODS.R
\name{login_irods}
\alias{login_irods}
\title{Login to iRods}
\usage{
login_irods(ttl = 1L, password = NULL)
}
\arguments{
\item{ttl}{Integer value indicating the hours the login has to be valid (time-to-live).}

\item{password}{Plain text password for irods, but never use literal string. See notes.}
}
\value{
TRUE if login succeeded, otherwise FALSE
}
\description{
Login to iRods
}
\note{
Never use a literal password string in a script.
Instead use a function that asks the user to enter a password or
a function that retrieves the password from a save location (e.g. keychain)
}
