% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_icommands.R
\name{login_yoda}
\alias{login_yoda}
\title{Inlog Yoda}
\usage{
login_yoda(ttl = 1L)
}
\arguments{
\item{ttl}{Integer value indicating the hours the login has to be valid (time-to-live)}
}
\description{
Inlog Yoda
}
