% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_iRODS.R
\name{set_irods_env}
\alias{set_irods_env}
\title{Set irods environment}
\usage{
set_irods_env(host = NULL, zone = NULL, user = NULL, override = FALSE)
}
\arguments{
\item{host}{Name (character string) of the organisation/department in data.uu.nl (e.g. 'i-lab.data.uu.nl')}

\item{zone}{Zone name (char string) of the irods organisation (e.g. 'nluu12p')}

\item{user}{Login name of user on irods host (e.g. j.jansen@uu.nl)}

\item{override}{Boolean (default FALSE) to indicate if it's allowed to override an existing irods environment}
}
\description{
Writes the irods host and user configuration items in HOME/.irods/irods_environment.json` file.
}
\note{
Currently very UU specific
}
